+++ armcmx/lpc11uxx/i2c.h
@@ -123,19 +123,19 @@ boolean i2c_receive(i2c * I2Cbuf, uint8_t * data, uint16_t lim);
 */
 
 void I2C_IRQHandler( void );
-uint8_t I2C_init(I2CDef * i2c, uint32 I2cMode );
-uint8_t I2C_begin(I2CDef * i2c);
+uint32 I2C_init(I2CDef * i2c, uint32 I2cMode );
+uint32 I2C_begin(I2CDef * i2c);
 
 uint32 I2C_start(I2CDef * i2c);
 uint32 I2C_stop(I2CDef * i2c);
 uint32 I2C_Engine(I2CDef * i2c);
 // returns false if issuing the start condition has been timed out.  
-uint8 I2C_write(I2CDef * i2c, uint8 addr, uint8 * data, size_t length);
-uint8 I2C_read(I2CDef * i2c, uint8 addr, uint8 * data, size_t reqlen, size_t reclen);
-uint8_t I2C_request(I2CDef * i2c, uint8_t addr, uint8_t * data, size_t reqlen);
-uint8_t I2C_receive(I2CDef * i2c, uint8_t addr, uint8_t * data, size_t reclen);
+uint32 I2C_write(I2CDef * i2c, uint8 addr, uint8 * data, uint32_t length);
+uint32 I2C_read(I2CDef * i2c, uint8 addr, uint8 * data, uint32_t reqlen, uint32_t reclen);
+uint32 I2C_request(I2CDef * i2c, uint8_t addr, uint8_t * data, uint32_t reqlen);
+uint32 I2C_receive(I2CDef * i2c, uint8_t addr, uint8_t * data, uint32_t reclen);
 //
-uint8 I2C_write16(I2CDef * i2c, uint8 addr, uint16 data);
+uint32 I2C_write16(I2CDef * i2c, uint8 addr, uint16 data);
 
 #ifdef __cplusplus
 }
